chore: rename mms5 to flexo.mms everywhere
