+++ Dockerfile
@@ -6,6 +6,6 @@ RUN ./gradlew --no-daemon installDist
 FROM openjdk:17.0.2-jdk-slim
 WORKDIR application
 RUN apt-get update && apt-get install -y procps
-COPY --from=build application/build/install/org.openmbee.mms5.auth-service/ .
-ENTRYPOINT ["./bin/org.openmbee.mms5.auth-service"]
+COPY --from=build application/build/install/org.openmbee.flexo.mms.auth-service/ .
+ENTRYPOINT ["./bin/org.openmbee.flexo.mms.auth-service"]
 EXPOSE 8080
