+++ README.md
@@ -1,5 +1,5 @@
-# MMS5 Auth Service
+# Flexo MMS Auth Service
 
-This service currently allows MMS5 Service clients to authenticate with a ldap server.
+This service currently allows Flexo MMS Service clients to authenticate with a ldap server.
 
-For usage, see [MMS 5 Docs](https://mms5-deployment-guide.readthedocs.io/en/latest/)
+For usage, see [Flexo MMS Docs](https://flexo-mms-deployment-guide.readthedocs.io/en/latest/)
