+++ src/main/resources/application.conf.example
@@ -5,7 +5,7 @@ ktor {
         port = ${?PORT}
     }
     application {
-        modules = [ org.openmbee.mms5.auth.ApplicationKt.module ]
+        modules = [ org.openmbee.flexo.mms.auth.ApplicationKt.module ]
     }
 }
 
@@ -14,7 +14,7 @@ jwt {
     domain = ${?JWT_DOMAIN}
     audience = "jwt-audience"
     audience = ${?JWT_AUDIENCE}
-    realm = "MMS5 Microservices"
+    realm = "Flexo MMS Microservices"
     realm = ${?JWT_REALM}
     secret = "test1234"
     secret = ${?JWT_SECRET}
